+++ yaralyzer/decoding/bytes_decoder.py
@@ -8,11 +8,14 @@ from collections import defaultdict
 from operator import attrgetter
 from typing import List, Optional
 
+from rich.align import Align
+from rich.console import Console, ConsoleOptions, NewLine, RenderResult
 from rich.panel import Panel
 from rich.table import Table
 from rich.text import Text
+from yaralyzer import bytes_match
 
-from yaralyzer.bytes_match import BytesMatch
+#from yaralyzer.bytes_match import BytesMatch
 from yaralyzer.config import YaralyzerConfig
 from yaralyzer.decoding.decoding_attempt import DecodingAttempt
 from yaralyzer.encoding_detection.character_encodings import ENCODING, ENCODINGS_TO_ATTEMPT
@@ -22,7 +25,6 @@ from yaralyzer.helpers.dict_helper import get_dict_key_by_value
 from yaralyzer.helpers.rich_text_helper import CENTER, DECODING_ERRORS_MSG, NO_DECODING_ERRORS_MSG
 from yaralyzer.output.decoding_attempts_table import (DecodingTableRow, assessment_only_row,
      build_decoding_attempts_table, decoding_table_row)
-from yaralyzer.output.rich_console import console
 from yaralyzer.util.logging import log
 
 # A 2-tuple that can be indexed by booleans of messages used in the table to show true vs. false
@@ -33,38 +35,47 @@ SCORE_SCALER = 100.0
 
 
 class BytesDecoder:
-    def __init__(self, bytes_match: BytesMatch, label: Optional[str] = None) -> None:
+    def __init__(self, bytes_match: 'BytesMatch', label: Optional[str] = None) -> None:
         self.bytes_match = bytes_match
         self.bytes = bytes_match.surrounding_bytes
         self.label = label or bytes_match.label
 
         # Empty table/metrics/etc
-        self.table = build_decoding_attempts_table(bytes_match)
         self.was_match_decodable = _build_encodings_metric_dict()
         self.was_match_force_decoded = _build_encodings_metric_dict()
         self.was_match_undecodable = _build_encodings_metric_dict()
         self.decoded_strings = {}  # dict[encoding: decoded string]
         self.undecoded_rows = []
+        self.decodings = []
 
         # Note we send both the match and surrounding bytes used when detecting the encoding
         self.encoding_detector = EncodingDetector(self.bytes)
 
-    def print_decode_attempts(self) -> None:
-        """Print the DecodingAttemptsTable."""
-        console.line(2)
-        self._print_decode_attempt_subheading()
+    def __rich_console__(self, _console: Console, options: ConsoleOptions) -> RenderResult:
+        """Rich object generator (see Rich console docs)"""
+        yield NewLine(2)
+        yield Align(self._decode_attempt_subheading(), CENTER)
 
-        if not YaralyzerConfig.SUPPRESS_CHARDET_OUTPUT:
-            console.line()
-            console.print(self.encoding_detector, justify='center')
-            console.line()
+        if not YaralyzerConfig.args.suppress_chardet:
+            yield NewLine()
+            yield Align(self.encoding_detector, CENTER)
+            yield NewLine()
 
-        console.print(self._generate_decodings_table())
+        # In standalone mode we always print the hex/raw bytes
+        if self.bytes_match.is_decodable():
+            yield self._generate_decodings_table()
+        elif YaralyzerConfig.args.standalone_mode:
+            # TODO: yield self.bytes_match.suppression_notice()
+            yield self._generate_decodings_table(True)
 
-    def _generate_decodings_table(self) -> Table:
-        """First rows are the raw / hex views of the bytes, then attempted decodings"""
-        if not self.bytes_match.is_decodable():
-            log.debug(f"{self.bytes_match} is not decodable")
+        yield NewLine()
+        yield Align(self.bytes_match.bytes_hashes_table(), CENTER, style='dim')
+
+    def _generate_decodings_table(self, suppress_decodes: bool = False) -> Table:
+        """First rows are the raw / hex views of the bytes, next rows are the attempted decodings"""
+        self.table = build_decoding_attempts_table(self.bytes_match)
+
+        if YaralyzerConfig.args.suppress_decoding_attempts or suppress_decodes:
             return self.table
 
         self.decodings = [
@@ -103,11 +114,10 @@ class BytesDecoder:
         """Check whether a given encoding is in the table already"""
         return any(row.encoding == encoding for row in self.decodings)
 
-    def _print_decode_attempt_subheading(self) -> None:
+    def _decode_attempt_subheading(self) -> Panel:
         """Generate a rich.Panel for decode attempts"""
         headline = Text(f"Found ", style='decode.subheading') + self.bytes_match.__rich__()
-        panel = Panel(headline, style='decode.subheading', expand=False)
-        console.print(panel, justify=CENTER)
+        return Panel(headline, style='decode.subheading', expand=False)
 
     def _track_decode_stats(self) -> None:
         """Track stats about successful vs. forced vs. failed decode attempts"""
@@ -147,7 +157,6 @@ class BytesDecoder:
 
         was_forced = WAS_DECODABLE_YES_NO[int(decoding.was_force_decoded)]
         return decoding_table_row(assessment, was_forced, display_text, sort_score)
-
 
 
 def _build_encodings_metric_dict():